We have an ok first go at a formated team report.
This does not yet include comments.

We have also not pursued what development of other report types for formats might entail.

We are considering the question as to what features we would want for a first demo.

Perhaps we have a sufficiently developed report structure for a demo.  Considering.  If so, we might turn toward developing an event framework.

We still have the question of what to do for a demo of the interactive service.